remove server property from the open-api configuration
Motivation:

The open-api for documenting cerebrum api allows setting the
server url property to the preferred url. In the initial state,
it was assumed that cerebrum will be served from localhost.
However, this is currently not the case, the application is
dockerised and can be served from different url than the assumed
one. Hence it is neccessary that we let the open-api generate
the server url.

Modification:

Remove server property from the open-api configuration.

Result:

The doc server url is now autogenrated.

Target: master
Review: https://gitlab.hzdr.de/hifis-technical-platform/helmholtz-cerebrum/-/merge_requests/35