fix uuid generation
Motivation:

A new uuid is generated for an entity that will be save on the db..
This is use as a primary way of uniquely identify a particular node.
To ensure the uniqueness of the uuid, two factor were considered
which are:
- uuid version 1
- secure random number
However the secured random number is currently generated causing the
thread to block in a VM (for details explaination see
https://tersesystems.com/blog/2015/12/17/the-right-way-to-use-securerandom/).
Hence the behaviour of the system become unpredicate and most likely
will fail if the request will required the uuid to be generated.

Modification:

- change how the secured random number is generated
- minor typo fix inside the MarketUserRepository

Result:

The generation of cerebrum custom uuid now work seemly on both local
machine and VMs. Hence, all requests are now process as expected.

Target: master