add env variable to dockerfile and adjust ci/cd deploy stage
Motivation:

Since we are moving to dockerize deployment, some variables,
like the neo4j hostname needs to be set different base on
the environment set up.

Also, we need to seperated the images that will be generated
when a normal commit is made and when a release is trigger.

Modification:

- expose the neo4j host name as configurable variable
- seperate docker image generated from a commit and a tag
    release
- add production_deployment to ci/cd deploy stage. This
   will be use to depoly to the production system. However,
   it just include a test to see if the ssh will work.

Result:

Two different docker images will be build, one for stable
tagged release and the other for development. These images
will be push to different docker repositories.

Target: master