add a new stage
Motivation:

Automated ci/cd is desirable process we will like to adhere to in 
our pipeline. Since we already have a local kubernetes cluster running 
on the production machine, we need to ensure that when a new tag
is release, the cluster is notified.

Modification:

- add a new pipeline stage called `docker`. This will be use for 
the building and deploy the image into the docker hub.
- deploy stage will now be responsible for trigger automated update 
of the image in the cluster.

Result:

Getting closer to fully automated ci/cd

Target: master