replace the neo4j with mongo db
Motivation:

We need to store media data for some of the entities and neo4j is
not design to store images, video, etc.. Hence, to avoid using
multiple dbs, we sort for a solution that can store the metadata
and the media data.

Modification:

- remove neo4j dependency and add mongo
- change the entities details and re-design it to suit the new db
- adjust the api design
- delete all neo4j relationship pojos
- adjust the tests according to the new definitions

Result:

Cerebrum now uses mongo db.

Target: master